[multiline] homepage architecture WIP
This commit begins a different architectural approach for the
(courses#index) by considering building it in React components.
Importantly, it also introduces a cleaner interface for the JBuilder,
removing the object instantiation and logic from the controller and
wrapping it in a presenter to more closely adhere to the Single
Responsibility Principle.